adding 5 padding between circles, introducing a starting point record to get all the numbers under control, changing for loop to foreach to get rid of two unnecessary numbers

+++ src/repo-man.domain/Diagram/SvgStringBuilder.cs
@@ -12,7 +12,7 @@ public class SvgStringBuilder
         _colorMapper = colorMapper;
     }
 
-    public void AddBoundingRectangle(int rectangleX, long rectangleY, long width,
+    public void AddBoundingRectangle(long rectangleX, long rectangleY, long width,
         long height, string folderName)
     {
         _builder.Append($"<g transform=\"translate({rectangleX},{rectangleY})\">");
